customer_service_assistant/requirements.txt
fastapi
httpx
uvicorn
loguru
pydantic
pyyaml
ollama
streamlit
watchdog
toml
zmq